requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
uvicorn[standard]
